requirements.txt
PyPDF2
python-dotenv
openai
docling
pytesseract
azure-cognitiveservices-vision-computervision
PyMuPDF